requirements.txt
Django>=3.0
psycopg2-binary>=2.8
djangorestframework
djoser
stream-chat
django-cors-headers
